Now question banks of a course are only accessed by the course instructors only

+++ src/main/java/com/lms/questionbank/QuestionBankController.java
@@ -1,10 +1,13 @@
 package com.lms.questionbank;
 
 
+import com.lms.course.CourseService;
 import com.lms.questionbank.question.mcq.MCQ;
 import com.lms.questionbank.question.Question;
 import com.lms.questionbank.question.QuestionDTO;
 import com.lms.questionbank.question.shortanswerquestion.ShortAnswerQuestion;
+import com.lms.user.User;
+import com.lms.user.UserService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,6 +15,7 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.security.Principal;
 import java.util.Collection;
 
 @RestController
@@ -20,38 +24,56 @@ import java.util.Collection;
 @AllArgsConstructor
 public class QuestionBankController {
     private final QuestionBankService questionBankService;
-
+    private final UserService userService;
+    private final CourseService courseService;
+    private boolean checkInstructor(Principal principal,@PathVariable long courseId) {
+        User currentInstructor = userService.getUser(principal);
+        if(courseService.getCourseById(courseId).isPresent())
+            return courseService.getCourseById(courseId).get().getInstructor().getId() == currentInstructor.getId();
+        return false;
+    }
 
     @GetMapping("/{questionBankId}/question/{id}")
-    public Question getQuestion(@PathVariable long questionBankId, @PathVariable long id)
+    public Question getQuestion(@PathVariable long questionBankId, @PathVariable long id, Principal principal,@PathVariable long courseId)
     {
+        if(!(this.checkInstructor(principal,courseId)))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+
         return questionBankService.getQuestion(id,questionBankId).orElseThrow(() ->new ResponseStatusException(HttpStatus.NOT_FOUND,"Question not found"));
     }
 
     @GetMapping("/{questionBankId}")
-    public Collection<Question> getAllQuestions(@PathVariable long questionBankId)
+    public Collection<Question> getAllQuestions(@PathVariable long questionBankId, Principal principal,@PathVariable long courseId)
     {
+        if(!(this.checkInstructor(principal,courseId)))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         return questionBankService.getAllQuestions(questionBankId);
     }
 
     @GetMapping("{questionBankId}/{discriminator}")
-    public Collection<Question> getAllQuestionsOfType(@PathVariable long questionBankId, @PathVariable String discriminator)
+    public Collection<Question> getAllQuestionsOfType(@PathVariable long questionBankId, @PathVariable String discriminator, Principal principal,@PathVariable long courseId)
     {
+        if(!(this.checkInstructor(principal,courseId)))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         return this.questionBankService.getAllQuestionsOfType(discriminator,questionBankId);
     }
 
 
     @PostMapping
-    public ResponseEntity<?> addQuestionBank(@RequestBody QuestionBank questionBank,@PathVariable long courseId)
+    public ResponseEntity<?> addQuestionBank(@RequestBody QuestionBank questionBank,@PathVariable long courseId, Principal principal)
     {
+        if(!(this.checkInstructor(principal,courseId)))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         this.questionBankService.newQuestionBank(questionBank,courseId);
         return ResponseEntity.ok().build();
     }
 
 
     @PostMapping("/{questionBankId}")
-    public ResponseEntity<?> addQuestion(@RequestBody QuestionDTO newQuestion, @PathVariable long questionBankId)
+    public ResponseEntity<?> addQuestion(@RequestBody QuestionDTO newQuestion, @PathVariable long questionBankId, Principal principal,@PathVariable long courseId)
     {
+        if(!(this.checkInstructor(principal,courseId)))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         Question question;
         if(newQuestion.getQuestionType().equals("mcq"))
             question = MCQ.builder()
@@ -73,8 +95,10 @@ public class QuestionBankController {
     }
 
     @DeleteMapping("/{questionBankId}/question/{id}")
-    public ResponseEntity<?> deleteQuestion(@PathVariable long questionBankId, @PathVariable long id)
+    public ResponseEntity<?> deleteQuestion(@PathVariable long questionBankId, @PathVariable long id, Principal principal,@PathVariable long courseId)
     {
+        if(!(this.checkInstructor(principal,courseId)))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         this.questionBankService.deleteQuestion(id,questionBankId);
         return ResponseEntity.ok().build();
     }
